WIP GhostSea
The idea of:

1. Implementing a "Ghost" data-structure with externally supplied
   GhostToken.
2. Implementing a stand-alone mirror by wrapping the "Ghost"
   data-structure into a GhostSea.

still seems sound to my mind.

Unfortunately, the compiler is quite reluctant to accept the code:

- From difficulties in handling Higher Ranked Lifetime Bounds (for<'a>).
- To codegen crashes on said bounds.

The current show-stopper is that it seems impossible for the closure
supplied to GhostSea::apply (and co) to embed a reference in its return
type.

This is likely due to the issue that R itself would need to be generic
over the lifetime for the signature to make sense.